milestone-0.3.39: partial space resolution, more.
Partial spaces now hold a reference to a selection of hmaps therein.
Previously there were to be a seperate list maintained at that
reference holding all hmaps currently in given partial.  This however
is a redundancy to the larger extent.  All hmaps in the set of lod
envelopes used per frame are to be checked for changes.  Any changed
hmaps in a given partial are sent to remote nodes thru the selection
given for that partial.

* src/transform.c (recycle): Changed.  Now expects a null terminated
list of hmap pointer references.
* src/hmap.h (attributes): attribs_t member bits becomes sign, and
member modifiers becomes mode.
* src/hmap.c (diag_selection): Added.
* src/attribs.c (search_vohspace): Added second btoggles_t arg and now
masks for both the sign bits (those that are tested every pass of the
state increment) and mode bits (tested less frequently) are passed.
Any bits in either mask that are present in the hmap, cause a match.
* src/generator.h (partial_space): Namechange.  partial_t member hmaps
changed to selection.
* src/generator.h (partial_space): Removed partial_t member
partial_hmaps, as the partial->selection list of hmap references is now
null terminated.
* src/generator.c (mk_partial): Changes.  partial->selection is now a
null terminated list.  Among other things this requires the list to have
an extra element added to selectf_a and selectf_b.
* src/generator.c (rm_partial): Removed function scope vars contents,
and maps, unnecessairy (was fumbling while trying new ways of using
select_t).
* src/generator.c (test_send_partial_changes): Temporary test added.
* src/generator.c (test_detach_all_partials): Temporary test added.
* src/generator.c (test_ls_partial): Name changed to diag_ls_partials.
As well, all partials are now listed.
* src/generator.c (diag_hmaps_in_partial): Added.
* src/generator.c (diag_partial_by_ordinal): Added.
* src/ifnodeglx.c (diag_node): Renamed to diag_node_key_f.  Now there
are 4 remappable functions for diagnostic use on keypress.  These are
then cascaded to where needed, for example generator.c
diag_generator_key_f.
* src/ifnodeglx.c (diag_node_key_g): Added.
* src/ifnodeglx.c (diag_node_key_h): Added.
* src/ifnodeglx.c (diag_node_key_i): Added.
* src/generator.c (diag_generator_key_f): Added.
* src/generator.c (diag_generator_key_g): Added.
* src/generator.c (diag_generator_key_h): Added.
* src/generator.c (diag_generator_key_j): Added.
* src/ifnodeglx.c (node): Temporary use of Esc key to exit program now
ends.  A helpfull message is (for now) output regarding this, informing
the person running vrtater of the new temporairy exit key.  Still not
sure what key to use, however the program will also exit from a menu
who's parent is spawned in ifnodeglx.c.
* src/session.c (publish_partial): Template code now prints the hmaps
that would be published to partial when called.  As well function scope
arg nodemap becomes maps.
* src/transform.c (surfinv_hmapf): Name changed to surface_inv_hmapf.